hackthebox/challenges/forensics/perseverance/sources/pyproject.toml
[tool.poetry]
name = "wmi"
version = "0.1.0"
description = ""
authors = ["apehex <apehex@protonmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
python-cim = '*'

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
